requirements.txt
flask==3.0.2
python-dotenv==1.0.0
Pillow==10.2.0
requests==2.31.0
werkzeug==3.0.1
pytest==8.0.0